+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to Swish will be documented here
 # 2.4 (In development)
 - Tapping a tab button will take you to its initial screen (#55)
 - Fix bug where the LeadTracker graph would not show overtime (#70)
+- Player name column in a boxscore is now fixed; only stats are now scrollable (#72)
 
 # 2.3 (5-6-2018)